[ticket/12609] Rename the interface in dispatcher_interface
PHPBB3-12609
