+++ phpBB/phpbb/event/dispatcher.php
@@ -13,21 +13,17 @@
 
 namespace phpbb\event;
 
+use Symfony\Component\EventDispatcher\ContainerAwareEventDispatcher;
+
 /**
-* Extension of the Symfony2 EventDispatcher
-*
-* It provides an additional `trigger_event` method, which
-* gives some syntactic sugar for dispatching events. Instead
-* of creating the event object, the method will do that for
-* you.
-*
-* Example:
-*
-*     $vars = array('page_title');
-*     extract($phpbb_dispatcher->trigger_event('core.index', compact($vars)));
-*
+* {@inherit}
 */
-interface dispatcher extends \Symfony\Component\EventDispatcher\EventDispatcherInterface
+class dispatcher extends ContainerAwareEventDispatcher implements dispatcher_interface
 {
-	public function trigger_event($eventName, $data = array());
+	public function trigger_event($eventName, $data = array())
+	{
+		$event = new \phpbb\event\data($data);
+		$this->dispatch($eventName, $event);
+		return $event->get_data_filtered(array_keys($data));
+	}
 }
